requirements.txt
streamlit
groq
pinecone-client
langchain
langchain-pinecone
sentence-transformers
plotly
yfinance
python-dotenv
openai
scikit-learn 